Bunch of changes  * Refactor numbering of bytecode (starts at 0 now)  * Fix stupid error in definition of BCMOp  * Add basic outline of vm.c  * Cleaned up documentation and changed operand info table structure

+++ bc.c
@@ -4,7 +4,7 @@
 
 /* Here we use BcMode + BytecodeDef to define the actual
    instruction encodings of our bytecode */
-uint32_t bc_encodings[] = {
+uint16_t bc_encodings[] = {
   /* Extra #define to append comma for enum so we can reuse BcMode */
 #define BcMode1(op, name, p1, p2, p3) BcMode(op,name,p1,p2,p3),
   BytecodeDef(BcMode1)
